requirements.txt
Django==3.2.3
mysqlclient==2.0.
pycryptodome==3.10.1
djangorestframework==3.12.4
django-fernet-fields==0.6
colorama==0.4.4
django-cors-headers==3.7.0